5.5 어린이날!!

+++ DX11 Framework/protocol.h
@@ -39,9 +39,9 @@ struct sc_packet_put_player {	//서버에서 처음 접속했을때 위치값과 ID를 부여한다.
 
 	WORD id;
 
-	int x;
-	int y;
-	int z;
+	float x;
+	float y;
+	float z;
 
 	BYTE	Animation;
